feat: delete docs by (`SegmentId`, `DocId`) (#26)
This teaches tantivy how to "directly" delete a document in a segment.

Our use case from pg_search is that we already know the segment_id and doc_id so it's waaaaay more efficient for us to delete docs through our `ambulkdelete()` routine.

It avoids doing a search, and all the stuff around that, for each of our "ctid" terms that we want to delete.