feat: ability to assign a panic handler to a Directory
Tantivy creates thread pools for some of its background work, including
committing and merging.

It's possible if one of the thread workers panics that rayon will simply
abort the process.  This is terrible from pg_search as that takes down
the entire Postgres cluster.

These changes allow a Directory to assign a panic handler that gets
called in such cases.  Which allows pg_search to gracefully rollback the
current transaction, while presenting the panic message to the user.