feat: Add `Directory::wants_cancel()` function
This adds a function named `wants_cancel() -> bool` to the `Directory`
trait.  It allows a Directory implementation to indicate that it would
like Tantivy to cancel an operation.

Right now, querying this function only happens during key points of
index merging, but _could_ be used in other places.  Technically,
segment merging is the only "black box" in tantivy that users don't
otherwise have the direct ability to control.

The default implementaiton of `wants_cancel()` returns false, so there's
no fear of default tantivy spuriously cancelling a merge.

The cancels happen "cleanly" such that if `wants_cancel()` returns true
an `Err(TantivyError::Cancelled)` is returned from the calling function
at that point, and the error result will be propogated up the stack.  No
panics are raised.