Fix `TopNComputer` threshold comparison (#51)
The `TopNComputer` was refactored in 0e04ec3136b9553f7435fd2ac11d2dc325d4cede to support both ascending and descending order. But that refactor did not adjust the `TopNComputer::threshold` to include the `ComparableDoc`, which is what provides the relevant `Ord` implementation.

Add a simple failing test, make it pass, and then expand the tests into proptests (which are already in use elsewhere in the codebase).

Upstream at: https://github.com/quickwit-oss/tantivy/pull/2651