perf: lazily open positions file
Not all queries use positions and it's okay if we (from the perspective
of `pg_search`, anyways) defer opening/loading them until they're first
needed.

This saves a lot of disk I/O when an index has lots of segments and the
query doesn't need positions.

As a drive by, make sure a random Vec has enough space before pushing
items to it.  This showed up in the profiler, believe it or not.