fix: Properly handle nulls in TopOrderable (#61)
Adjusts `Feature` to use `Option` where necessary. `StringFeature::SegmentOutput` avoids being wrapped in an `Option` because the `u64::MAX` "niche" is safe to use in that case (since it would require that many terms to trigger a collision).

Surprisingly, no performance impact downstream.

I additionally needed to adjust the unit tests to:
* include a NULL
* use a stable id, because for some reason at three segments (but not two?) the fact that segment ords are not stable was exposed.